msm8996-common: bust partition probing for two seconds
It takes aprox. 2s to get the recovery to detect the new "vendor" partition
(in case it's added to TWRP mount flags), so hold it for that time until
we can check the path later on.

adb shell dmesg | grep sde34
<6>[    1.596729]  sde: sde1 sde2 sde3 sde4 sde5 sde6 sde7 sde8 sde9 sde10 sde11 sde12 sde13 sde14 sde15 sde16 sde17 sde18 sde19 sde20 sde21 sde22 sde23 sde24 sde25 sde26 sde27 sde28 sde29 sde30 sde31 sde32 sde33 sde34
<6>[   30.522532]  sde: sde1 sde2 sde3 sde4 sde5 sde6 sde7 sde8 sde9 sde10 sde11 sde12 sde13 sde14 sde15 sde16 sde17 sde18 sde19 sde20 sde21 sde22 sde23 sde24 sde25 sde26 sde27 sde28 sde29 sde30 sde31 sde32 sde33 sde34
<6>[   31.046327] EXT4-fs (sde34): mounted filesystem with ordered data mode. Opts:
<7>[   31.046353] SELinux: initialized (dev sde34, type ext4), uses xattr

It's less than a second actually, but we provide a failsafe 2s hold
here.

Change-Id: I8a0a66c7b492deba2e40240f16ac3f69f5908a11
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>